MVP (#72)
* MVP and tests

+++ scripts/tests.e2e.sh
@@ -1,23 +1,39 @@
 #!/usr/bin/env bash
 set -e
 
-# ./scripts/tests.e2e.sh 1.7.11
+# build spacesvm-rs binary
+# ./scripts/build.release.sh
+#
+# download from github, keep network running
+# VM_PLUGIN_PATH=$(pwd)/target/release/spacesvm-rs ./scripts/tests.e2e.sh
+#
+# download from github, shut down network
+# NETWORK_RUNNER_ENABLE_SHUTDOWN=1 VM_PLUGIN_PATH=$(pwd)/target/release/spacesvm-rs ./scripts/tests.e2e.sh
+#
+# use custom avalanchego binary
+# VM_PLUGIN_PATH=$(pwd)/target/release/timestampvm ./scripts/tests.e2e.sh ~/go/src/github.com/ava-labs/avalanchego/build/avalanchego
+#
 if ! [[ "$0" =~ scripts/tests.e2e.sh ]]; then
   echo "must be run from repository root"
   exit 255
 fi
 
+AVALANCHEGO_PATH=${1:-""}
+echo AVALANCHEGO_PATH: ${AVALANCHEGO_PATH}
+echo VM_PLUGIN_PATH: ${VM_PLUGIN_PATH}
+
 #################################
 # download avalanche-network-runner
 # https://github.com/ava-labs/avalanche-network-runner
 # TODO: use "go install -v github.com/ava-labs/avalanche-network-runner/cmd/avalanche-network-runner@v${NETWORK_RUNNER_VERSION}"
-GOOS=$(go env GOOS) # ensures that the compatible network runner version is downloaded for this machine
-NETWORK_RUNNER_VERSION=1.1.0
+GOOS=$(go env GOOS)
+NETWORK_RUNNER_VERSION=1.3.0
 DOWNLOAD_PATH=/tmp/avalanche-network-runner.tar.gz
 DOWNLOAD_URL=https://github.com/ava-labs/avalanche-network-runner/releases/download/v${NETWORK_RUNNER_VERSION}/avalanche-network-runner_${NETWORK_RUNNER_VERSION}_linux_amd64.tar.gz
 if [[ ${GOOS} == "darwin" ]]; then
   DOWNLOAD_URL=https://github.com/ava-labs/avalanche-network-runner/releases/download/v${NETWORK_RUNNER_VERSION}/avalanche-network-runner_${NETWORK_RUNNER_VERSION}_darwin_amd64.tar.gz
 fi
+echo ${DOWNLOAD_URL}
 
 rm -f ${DOWNLOAD_PATH}
 rm -f /tmp/avalanche-network-runner
@@ -36,20 +52,37 @@ echo "launch avalanche-network-runner in the background"
 server \
 --log-level debug \
 --port=":12342" \
---grpc-gateway-port=":12343" &
+--disable-grpc-gateway &
 NETWORK_RUNNER_PID=${!}
-sleep 5 # sleep to ensure that network runner initializes before e2e tests begin
+sleep 5
 
 #################################
-# do not run in parallel, to run in sequence
 echo "running e2e tests"
 NETWORK_RUNNER_GRPC_ENDPOINT=http://127.0.0.1:12342 \
+AVALANCHEGO_PATH=${AVALANCHEGO_PATH} \
+VM_PLUGIN_PATH=${VM_PLUGIN_PATH} \
 RUST_LOG=debug \
 cargo test --all-features --package e2e -- --show-output --nocapture
 
 #################################
-# "e2e.test" already terminates the cluster for "test" mode
-# just in case tests are aborted, manually terminate them again
-echo "network-runner RPC server was running on NETWORK_RUNNER_PID ${NETWORK_RUNNER_PID} as test mode; terminating the process..."
-pkill -P ${NETWORK_RUNNER_PID} || true
-kill -2 ${NETWORK_RUNNER_PID}
+echo ""
+echo ""
+if [ -z "$NETWORK_RUNNER_ENABLE_SHUTDOWN" ]
+then
+  echo "SKIPPED SHUTDOWN..."
+  echo ""
+  echo "RUN FOLLOWING TO CLEAN UP:"
+  echo "pkill -P ${NETWORK_RUNNER_PID} || true"
+  echo "kill -2 ${NETWORK_RUNNER_PID} || true"
+  echo ""
+else 
+  echo "SHUTTING DOWN..."
+  echo ""
+  # "e2e.test" already terminates the cluster for "test" mode
+  # just in case tests are aborted, manually terminate them again
+  echo "network-runner RPC server was running on NETWORK_RUNNER_PID ${NETWORK_RUNNER_PID} as test mode; terminating the process..."
+  pkill -P ${NETWORK_RUNNER_PID} || true
+  kill -2 ${NETWORK_RUNNER_PID} || true
+fi
+
+echo "TEST SUCCESS"
